Added some new sound FX
The chaos crumbles are little glitch sounds edited from a raw session on VCV rack 2. I left in the reaper session and the raw files where I edited them, but my suggested sounds are the files CC01-CC10.wav. All 44.1khz 24 bit. Let me know if we need to dither these down to 16 bit. What might be a good idea is to have a general "chaotic sound" toggle in the program, where it picks a random number from 1-10 and plays the corresponding file